+++ lib/unisec/cli/cli.rb
@@ -33,9 +33,11 @@ module Unisec
       register 'dump codepoints standard', Dump::Codepoints::Standard
       register 'dump dec', Dump::Dec
       register 'dump hex', Dump::Hex
+      register 'dump rev', Dump::Reverse
       register 'grep', Grep
       register 'normalize all', Normalize::All
       register 'normalize replace', Normalize::Replace
+      register 'normalize reverse', Normalize::Reverse
       register 'planes list', Planes::List
       register 'planes search', Planes::Search
       register 'properties char', Properties::Char