+++ lib/unisec/normalization.rb
@@ -95,6 +95,35 @@ module Unisec
       Normalization.replace_bypass(@original)
     end
 
+    # Find the list of symbols that will transform into a given symbol after normalization
+    # @param target [String]
+    # @param forms [String|Symbol|Array<Symbol>]
+    # @return [Hash] (results won't include input)
+    # @example
+    #   Unisec::Normalization.reverse_normalize('<') # => {nfc: [], nfd: [], nfkc: ["﹤", "＜"], nfkd: ["﹤", "＜"]}
+    #   Unisec::Normalization.reverse_normalize('.', forms: [:nfkc, :nfkd]) # => {nfkc: ["․", "﹒", "．"], nfkd: ["․", "﹒", "．"]}
+    #   Unisec::Normalization.reverse_normalize('ffi', forms: :nfkc) # => {nfkc: ["ﬃ"]}
+    #   Unisec::Normalization.reverse_normalize('≯', forms: 'nfd') # => {nfd: ["≯"]}
+    #   Unisec::Normalization.reverse_normalize('ô', forms: 'nfc,nfd') # => {nfc: [], nfd: []}
+    def self.reverse_normalize(target, forms: %i[nfc nfd nfkc nfkd])
+      forms = Utils::Arguments.to_array_of_sym(forms)
+      result = {}
+      forms.each do |form|
+        result[form] = []
+      end
+
+      (0x000000..0x10FFFF).each do |codepoint|
+        char = codepoint.chr(Encoding::UTF_8)
+        forms.each do |form|
+          result[form] << char if (char.unicode_normalize(form) == target) && (char != target)
+        end
+      rescue RangeError # skip UTF-16 surrogates and potential other invalid code points
+        next
+      end
+
+      result
+    end
+
     # Display a CLI-friendly output summurizing all normalization forms
     # @return [String] CLI-ready output
     # @example
@@ -124,6 +153,18 @@ module Unisec
 
     # Display a CLI-friendly output of the XSS payload to bypass HTML escape and
     # what it does once normalized in NFKC & NFKD.
+    # @return [String] CLI-ready output
+    # @example
+    #   $ puts Unisec::Normalization.new('<script>').display_replace
+    #   # =>
+    #   # Original: <script>
+    #   #   U+003C U+0073 U+0063 U+0072 U+0069 U+0070 U+0074 U+003E
+    #   # Bypass payload: ＜script＞
+    #   #   U+FF1C U+0073 U+0063 U+0072 U+0069 U+0070 U+0074 U+FF1E
+    #   # NFKC: <script>
+    #   #   U+003C U+0073 U+0063 U+0072 U+0069 U+0070 U+0074 U+003E
+    #   # NFKD: <script>
+    #   #   U+003C U+0073 U+0063 U+0072 U+0069 U+0070 U+0074 U+003E
     def display_replace
       colorize = lambda { |form_title, form_attr|
         "#{Paint[form_title.to_s, :underline,
@@ -134,6 +175,37 @@ module Unisec
         colorize.call('Bypass payload', payload) +
         colorize.call('NFKC', Normalization.nfkc(payload)) +
         colorize.call('NFKD', Normalization.nfkd(payload))
+    end
+
+    # Display a CLI-friendly output reverse normalization results
+    # @param target [String] see {Unisec::Normalization.reverse_normalize}
+    # @param forms [String|Symbol|Array<Symbol>] see {Unisec::Normalization.reverse_normalize}
+    # @return [String] CLI-ready output
+    # @example
+    #   puts Unisec::Normalization.display_reverse_normalize('<')
+    #   # =>
+    #   # Original:
+    #   #   < (U+003C)
+    #   # NFKC
+    #   #   ﹤ (U+FE64)
+    #   #   ＜ (U+FF1C)
+    #   # NFKD
+    #   #   ﹤ (U+FE64)
+    #   #   ＜ (U+FF1C)
+    def self.display_reverse_normalize(target, forms: %i[nfc nfd nfkc nfkd]) # rubocop:disable Metrics/AbcSize
+      colorize_form = ->(form_title) { Paint[form_title, :underline, :bold] }
+      colorize_char = ->(char) { "  #{char} (#{Paint[Unisec::Utils::String.chars2codepoints(char), :red]})\n" }
+      out = "#{colorize_form.call('Original')}:\n#{colorize_char.call(target)}"
+      res = Unisec::Normalization.reverse_normalize(target, forms: forms) # => {nfc: [], nfd: [], nfkc: ["﹤", "＜"], nfkd: ["﹤", "＜"]}
+      res.each_key do |k|
+        next if res[k].empty?
+
+        out += "#{colorize_form.call(k.to_s.upcase)}\n"
+        res[k].each do |v|
+          out += colorize_char.call(v)
+        end
+      end
+      out
     end
   end
 end