+++ test/test_utils.rb
@@ -67,4 +67,16 @@ class UnisecTest < Minitest::Test
     assert_equal(false, (1..10).include_range?(2..11))
     assert_equal(true, (1..10).include_range?(2..4))
   end
+
+  def test_unisec_utils_arguments_to_array_of_sym
+    assert_equal([:arg], Unisec::Utils::Arguments.to_array_of_sym("arg"))
+    assert_equal([:a, :b, :c], Unisec::Utils::Arguments.to_array_of_sym("a,b,c"))
+    assert_equal([:snake], Unisec::Utils::Arguments.to_array_of_sym(:snake))
+    assert_equal([:a, :b, :c], Unisec::Utils::Arguments.to_array_of_sym([:a, :b, :c]))
+  end
+
+  def test_unisec_utils_arguments_argenc2enc
+    assert_equal('UTF-8', Unisec::Utils::Arguments.argenc2enc('utf8', target: 'standard'))
+    assert_equal(Encoding::UTF_16BE, Unisec::Utils::Arguments.argenc2enc('utf16be', target: 'class'))
+  end
 end